Runtime: Module::DeclareAndBindNativeFunction returns the index of the bound function.

+++ include/pulsar/runtime.h
@@ -127,8 +127,10 @@ namespace Pulsar
         RuntimeState ExecuteFunction(const FunctionDefinition& def, ValueStack& stack, ExecutionContext& context) const;
 
         typedef std::function<RuntimeState(ExecutionContext&)> NativeFunction;
+        // Returns how many definitions were bound.
         size_t BindNativeFunction(const FunctionDefinition& def, NativeFunction func);
-        size_t DeclareAndBindNativeFunction(FunctionDefinition def, NativeFunction func);
+        // Returns the index of the newly declared function.
+        int64_t DeclareAndBindNativeFunction(FunctionDefinition def, NativeFunction func);
         uint64_t BindCustomType(const String& name, CustomType::DataFactory_T dataFactory = nullptr);
 
         // Be sure to check if the type exists first (unless you know for sure it exists)
